refactor: reorganize codebase into hierarchical module structure
Transform flat module organization into structured hierarchy:
- Commands: Extract individual command handlers to src/commands/
- Core: Separate git operations (src/core/git/) from runtime (src/core/runtime/)
- Logging: Consolidate logging functionality in src/logging/
- App: Add application context management in src/app/
- Update all import paths and test references to match new structure

This architectural change improves code organization and maintainability
while preserving all existing functionality.
