+++ src/core/git/mod.rs
@@ -1,0 +1,14 @@
+pub mod diff;
+pub mod gc;
+pub mod index;
+pub mod repo;
+pub mod restore;
+pub mod shell;
+pub mod snapshot;
+
+pub use diff::{DiffFormat, diff};
+pub use gc::gc;
+pub use repo::{autosnap_dir, init_autosnap, repo_root};
+pub use restore::restore;
+pub use shell::snapshot_shell;
+pub use snapshot::snapshot_once;
