+++ src/logging/mod.rs
@@ -1,0 +1,2 @@
+pub mod init;
+pub mod view;
