+++ src/logging/view.rs
