+++ src/main.rs
@@ -1,7 +1,9 @@
 use anyhow::Result;
 use clap::Parser;
 use git_autosnap::cli::{Cli, Commands};
-use git_autosnap::{gitlayer, init_tracing, init_tracing_with_file, run};
+use git_autosnap::core::git;
+use git_autosnap::logging::init::{init_tracing, init_tracing_with_file};
+use git_autosnap::run;
 
 fn main() -> Result<()> {
     let cli = Cli::parse();
@@ -10,7 +12,7 @@ fn main() -> Result<()> {
     // Start command needs file logging for the watcher
     match &cli.command {
         Commands::Start { daemon } => {
-            let root = gitlayer::repo_root()?;
+            let root = git::repo_root()?;
             init_tracing_with_file(&root, cli.verbose, *daemon)?;
         }
         _ => {
